post

+++ app/Http/Controllers/TasksController.php
@@ -16,14 +16,21 @@ class TasksController extends Controller
     public function index()
     {
         
-        $tasks = Task::all();
+         $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+             ];
     }
-
+        return view('welcome', $data);
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -56,6 +63,7 @@ class TasksController extends Controller
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::id();
         $task->save();
 
         // トップページへリダイレクトさせる
@@ -129,12 +137,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-           // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 }
